feat: add a display error for search results
When an error was originated from the network, the app was staying in a state of infinite fetch mode. Now, if an error occurs from a request, the application will notify the user about it and will stop the current search. This way, the app can recover itself and continue functional. Issue: #39